[YUNIKORN-153] Move updateRequest creation out of resource.go
